requirements.txt
code128==0.3
Pillow==10.4.0
